Disable an assertion with valgrind builds.
There is an assertion that truncating a table with a view can fail when
the result of the view update may violate a constraint.  With normal
debug and release builds this is just right, since we delete row by row.
But for valgrind builds we delete all rows at once, which means the
views are not updated until the source tables are all truncated.  So the
constraint violations are not triggered.  Sicne the test was looking for
a violation the test failed under memcheck (== valgrind) builds.

+++ tests/frontend/org/voltdb/regressionsuites/TestGroupByComplexMaterializedViewSuite.java
@@ -1509,9 +1509,13 @@ public class TestGroupByComplexMaterializedViewSuite extends RegressionSuite {
 
         // With view size limits in place, it's actually not possible
         // to delete all the rows from the source table
-        verifyStmtFails(client, "truncate table r6", "exceeds the size");
-        validateTableOfScalarLongs(client, "select count(*) from (select id from r6 where id > -1000) as sel",
-                new long[] {numSourceRows});
+        if (! isValgrind()) {
+            // This does not actually fail under valgrind, where we
+            // allocate memory differently.
+            verifyStmtFails(client, "truncate table r6", "exceeds the size");
+            validateTableOfScalarLongs(client, "select count(*) from (select id from r6 where id > -1000) as sel",
+                    new long[] {numSourceRows});
+        }
     }
 
     //
